bug fixes, test changes
