+++ src/test/java/com/kttdevelopment/myanimelist/AnimeTests/TestUserAnimeListing.java
@@ -1,0 +1,39 @@
+package com.kttdevelopment.myanimelist.AnimeTests;
+
+import com.kttdevelopment.myanimelist.MyAnimeList;
+import com.kttdevelopment.myanimelist.TestProvider;
+import com.kttdevelopment.myanimelist.anime.AnimeListStatus;
+import com.kttdevelopment.myanimelist.anime.property.AnimeSort;
+import com.kttdevelopment.myanimelist.anime.property.AnimeStatus;
+import org.junit.jupiter.api.*;
+
+import java.util.List;
+
+public class TestUserAnimeListing {
+
+    private static MyAnimeList mal;
+
+    @BeforeAll
+    public static void beforeAll(){
+        mal = TestProvider.getMyAnimeList();
+    }
+
+    @Test
+    public void testStatus(){
+        final List<AnimeListStatus> list =
+            mal.getUserAnimeListing()
+               .withStatus(AnimeStatus.Dropped)
+               .search();
+        Assertions.assertEquals(AnimeStatus.Dropped, list.get(0).getStatus());
+    }
+
+    @Test
+    public void testSort(){
+        final List<AnimeListStatus> list =
+            mal.getUserAnimeListing()
+               .sortBy(AnimeSort.UpdatedAt)
+               .search();
+        Assertions.assertTrue(list.get(0).getUpdatedAt() > list.get(1).getUpdatedAt());
+    }
+
+}
